Move ShipAICache to the main source directory.
