+++ source/ShipAICache.cpp
@@ -15,9 +15,9 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "ShipAICache.h"
 
-#include "../Outfit.h"
-#include "../pi.h"
-#include "../Ship.h"
+#include "Outfit.h"
+#include "pi.h"
+#include "Ship.h"
 
 #include <algorithm>
 #include <cmath>
